Refactorized GildedRose class, rules for update now are in another file
Separate the class, now the rules for update items are in file.
utility_functions, the method update_quality check the items and call a
respective function
